Integrate code review suggestions
