+++ src/GitVersion.Core/Configuration/EffectiveBranchConfiguration.cs
@@ -1,5 +1,4 @@
 using GitVersion.Extensions;
-using GitVersion.VersionCalculation;
 
 namespace GitVersion.Configuration;
 
@@ -8,12 +7,4 @@ public record EffectiveBranchConfiguration(EffectiveConfiguration Value, IBranch
     public IBranch Branch { get; } = Branch.NotNull();
 
     public EffectiveConfiguration Value { get; } = Value.NotNull();
-
-    public NextVersion CreateNextVersion(BaseVersion baseVersion, SemanticVersion incrementedVersion)
-    {
-        incrementedVersion.NotNull();
-        baseVersion.NotNull();
-
-        return new NextVersion(incrementedVersion, baseVersion, this);
-    }
 }
