+++ src/GitVersion.Core/Core/IBranchRepository.cs
@@ -1,0 +1,8 @@
+namespace GitVersion.Core;
+
+internal interface IBranchRepository
+{
+    IEnumerable<IBranch> GetMainlineBranches(params IBranch[] excludeBranches);
+
+    IEnumerable<IBranch> GetReleaseBranches(params IBranch[] excludeBranches);
+}
